# src/rest-server/docs/swagger.yaml
openapi: 3.0.1
info:
  title: OpenPAI RESTful API
  description: 'Open Platform for AI RESTful API docs.'
  license:
    name: MIT License
    url: https://github.com/microsoft/pai/blob/master/LICENSE
  version: 0.16.0
...
paths:
...
  /api/v2/user:
    post:
      tags:
      - user
      summary: Create a user in the system.
      description: Create a user in the system by admin, basic authentication mode only.
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        description: User account
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: username in [\w.-]+ format
                password:
                  type: string
                  description: password at least 6 characters
                admin:
                  type: boolean
                email:
                  type: string
                  description: email address or empty string
                virtualCluster:
                  type: array
                  items:
                    type: string
                  description: vc list
                extension:
                  type: object
              required: [username, password]
        required: true
      responses:
        201:
          description: Succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: User is created successfully
        401:
          $ref: '#/components/responses/UnauthorizedUserError'
        403:
          $ref: '#/components/responses/ForbiddenUserError'
        409:
          $ref: '#/components/responses/ConflictUserError'
        500:
          $ref: '#/components/responses/UnknownError'
  /api/v2/user/{user}:
    delete:
      tags:
      - user
      summary: Remove a user in the system.
      description: Remove a user in the system. by admin, basic authentication mode only.
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/user'
      responses:
        200:
          description: Succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: user is removed successfully
        401:
          $ref: '#/components/responses/UnauthorizedUserError'
        403:
          $ref: '#/components/responses/ForbiddenUserError'
        404:
          $ref: '#/components/responses/NoUserError'
        500:
          $ref: '#/components/responses/UnknownError'
  /api/v2/user/{user}/password:
    put:
      tags:
      - user
      summary: Update user password.
      description: Administrator change other user's password; user can change his own password.
      operationId: updateUserPassword
      security:
        - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/user'
      requestBody:
        description: Old and new password
        content:
          application/json:
            schema:
              type: object
              properties:
                oldPassword:
                  type: string
                  description: admin could ignore this param
                newPassword:
                  type: string
                  description: password at least 6 characters
              required: [newPassword]
        required: true
      responses:
        201:
          description: Succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: update user password successfully.
        401:
          $ref: '#/components/responses/UnauthorizedUserError'
        403:
          $ref: '#/components/responses/ForbiddenUserError'
        404:
          $ref: '#/components/responses/NoUserError'
        500:
          $ref: '#/components/responses/UnknownError'
  /api/v2/user/{user}/virtualcluster:
    put:
      tags:
      - user
      summary: Update user virtual cluster.
      description: Administrator change other user's virtualCluster list.
      operationId: updateUserVirtualCluster
      security:
        - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/user'
      requestBody:
        description: Virtual clusters
        content:
          application/json:
            schema:
              type: object
              properties:
                virtualCluster:
                  type: array
                  items:
                    type: string
                  description: virtual cluster list
              required: [virtualCluster]
        required: true
      responses:
        201:
          description: Succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: Update user virtualCluster data successfully.
        400:
          $ref: '#/components/responses/NoVirtualClusterError'
        401:
          $ref: '#/components/responses/UnauthorizedUserError'
        403:
          $ref: '#/components/responses/ForbiddenUserError'
        404:
          $ref: '#/components/responses/NoUserError'
        500:
          $ref: '#/components/responses/UnknownError'
  /api/v2/user/{user}/email:
    put:
      tags:
      - user
      summary: Update user email.
      description: Administrator change other user's email address, and user could update his own email address.
      operationId: updateUserEmail
      security:
        - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/user'
      requestBody:
        description: Email
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: email address
              required: [email]
        required: true
      responses:
        201:
          description: Succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: Update user email data successfully.
        401:
          $ref: '#/components/responses/UnauthorizedUserError'
        403:
          $ref: '#/components/responses/ForbiddenUserError'
        404:
          $ref: '#/components/responses/NoUserError'
...
  /api/v2/user/{user}/extension:
    put:
      tags:
      - user
      summary: Update user extension.
      description: Administrator change other user's extension, and user could update his own extension.
      operationId: updateUserExtension
      security:
        - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/user'
      requestBody:
        description: Extension
        content:
          application/json:
            schema:
              type: object
              properties:
                extension:
                  type: object
              required: [extension]
        required: true
      responses:
        201:
          description: Succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: Update user extension data successfully.
        401:
          $ref: '#/components/responses/UnauthorizedUserError'
        403:
          $ref: '#/components/responses/ForbiddenUserError'
        404:
          $ref: '#/components/responses/NoUserError'
  /api/v2/user/{user}/grouplist:
    put:
      tags:
      - user
      summary: Update user grouplist.
      description: Administrator change other user's grouplist.
      operationId: updateUserGrouplist
      security:
        - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/user'
      requestBody:
        description: Group list
        content:
          application/json:
            schema:
              type: object
              properties:
                grouplist:
                  type: array
                  items:
                    type: string
                  description: group list
              required: [grouplist]
        required: true
      responses:
        201:
          description: Succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: update user grouplist successfully.
        401:
          $ref: '#/components/responses/UnauthorizedUserError'
        403:
          $ref: '#/components/responses/ForbiddenUserError'
        404:
          $ref: '#/components/responses/NoUserError'
...
components:
  parameters:
...
    user:
      name: user
      in: path
      description: user name
      required: true
      schema:
        type: string
...
  schemas:
    Response:
      type: object
      properties:
        code:
          type: string
          enum:
          - BadConfigurationError
          - ConflictJobError
          - ConflictUserError
          - ConflictVcError
          - ForbiddenUserError
          - ForbiddenKeyError
          - IncorrectPasswordError
          - InvalidParametersError
          - NoApiError
          - NoJobError
          - NoJobConfigError
          - NoJobSshInfoError
          - NoUserError
          - NoGroupError
          - NoVirtualClusterError
          - ReadOnlyJobError
          - RemoveAdminError
          - ReadOnlyVcError
          - RemoveRunningVcError
          - UnauthorizedUserError
          - NoEnoughQuotaError
          - NotImplementedError
          - UnknownError
        message:
          type: string
      required: [message]
...
  responses:
...
    NoVirtualClusterError:
      description: NoVirtualClusterError
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
          examples:
            NoVirtualClusterError:
              value:
                code: NoVirtualClusterError
                message: Virtual cluster {vc} is not found.
...
    UnauthorizedUserError:
      description: UnauthorizedUserError
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
          examples:
            UnauthorizedUserError:
              value:
                code: UnauthorizedUserError
                message: Guest is not allowed to do this operation.
    ForbiddenUserError:
      description: ForbiddenUserError
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
          examples:
            ForbiddenUserError:
              value:
                code: ForbiddenUserError
                message: Non-admin is not allow to do this operation.
...
    NoUserError:
      description: NoUserError
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
          examples:
            NoUserError:
              value:
                code: NoUserError
                message: User {user} is not found.
...
    ConflictUserError:
      description: ConflictUserError
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
          examples:
            ConflictUserError:
              value:
                code: ConflictUserError
                message: User name {user} already exists.
...
    UnknownError:
      description: UnknownError
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response'
          examples:
            UnknownError:
              value:
                code: UnknownError
                message: Internal error message.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT